refactor: add ge0 port interface
