--- Makefile
@@ -1,3 +1,0 @@
-NAME = game_engine
-SRCS = main.c screen.c display.c
-include $(AM_HOME)/Makefile
