+++ display.c
@@ -1,8 +1,6 @@
 #include "display.h"
-#include "amdev.h"
-#include "klib-macros.h"
-#include "klib.h"
 #include "screen.h"
+#include "ge0_port_interface.h"
 #include <stdbool.h>
 
 // This file includes actual logic to draw on hardware
@@ -11,36 +9,17 @@ uint16_t rscreenWidth, rscreenHeight; // Real screen size
 uint16_t displayXOffset;              // Margin of the display area
 uint32_t pix_buffer[SCREEN_REAL_WIDTH];
 
-static uint32_t rgb565_to_rgb888(uint16_t rgb565) {
-    // See https://stackoverflow.com/a/2445096
-    uint8_t r = (rgb565 >> 11) & 0x1F;
-    uint8_t g = (rgb565 >> 5) & 0x3F;
-    uint8_t b = rgb565 & 0x1F;
-    r = (r << 3) | (r >> 2);
-    g = (g << 2) | (g >> 4);
-    b = (b << 3) | (b >> 2);
-    return (r << 16) | (g << 8) | b;
-}
-void am_display_fillScreen(uint16_t color) { // engine uses 16 bit color(rgb565)
-    //  todo: read those two only one time
-    int height = io_read(AM_GPU_CONFIG).height;
-    int width = io_read(AM_GPU_CONFIG).width;
-    /* uint32_t empty_buffer[height * width]; */ // 爆栈了！
-    uint32_t *empty_buffer = malloc(height * width * sizeof(uint32_t));
-    uint32_t color_888 = rgb565_to_rgb888(color);
-    for (int i = 0; i < height * width; i++) {
-        /* printf("Current i is %d\n",i); */
-        empty_buffer[i] = color_888;
-    }
-    io_write(AM_GPU_FBDRAW, 0, 0, empty_buffer, width, height, false);
-    io_write(AM_GPU_FBDRAW, 0, 0, NULL, 0, 0, true);
-    free(empty_buffer);
-}
-void am_display_drawLine(uint32_t line, uint32_t start, uint32_t width,
-                         uint32_t *buf) {
-    io_write(AM_GPU_FBDRAW, start, line, buf, width, 1, 0);
-}
-void am_display_sync() { io_write(AM_GPU_FBDRAW, 0, 0, NULL, 0, 0, 1); }
+// 在GE0-YSYX中转换，ge0_port_interface接口应当统一使用RGB565颜色格式
+// static uint32_t rgb565_to_rgb888(uint16_t rgb565) {
+//     // See https://stackoverflow.com/a/2445096
+//     uint8_t r = (rgb565 >> 11) & 0x1F;
+//     uint8_t g = (rgb565 >> 5) & 0x3F;
+//     uint8_t b = rgb565 & 0x1F;
+//     r = (r << 3) | (r >> 2);
+//     g = (g << 2) | (g >> 4);
+//     b = (b << 3) | (b >> 2);
+//     return (r << 16) | (g << 8) | b;
+// }
 
 void setScreenResolution(uint16_t nw, uint16_t nh) {
     if (nw < SCREEN_REAL_WIDTH)
@@ -105,15 +84,15 @@ void redrawScreen() { // todo: change the name to display. "screen" is for
                                          0xf0) >>
                                         4];
                     }
-                    pix_buffer[x_physical] =
-                        rgb565_to_rgb888(pix_buffer[x_physical]);
+                    // 应该在GE0-YSYX中进行转换，ge0_port_interface接口应当统一使用RGB565颜色格式
+                    // pix_buffer[x_physical] =
+                    //     rgb565_to_rgb888(pix_buffer[x_physical]);
                 }
             prev_y_canvas = y_canvas;
             am_display_drawLine(y_physical, startx, rscreenWidth, pix_buffer);
         }
     }
-    // todo remove this
-    io_write(AM_GPU_FBDRAW, 0, 0, NULL, 0, 0, true);
+    am_display_sync();
     for (uint16_t i = 0; i < 4; i++)
         line_is_draw[i] = 0;
 }
