+++ include/display.h
@@ -1,8 +1,8 @@
 #ifndef __DISPLAY_H__
 #define __DISPLAY_H__
 
-#include <am.h>
-#include <klib-macros.h>
+#include <stdint.h>
+
 // todo: change this two
 // this two in orig project keep the real hw size of the screen
 // while the render size is configurable
@@ -10,5 +10,6 @@
 #define SCREEN_REAL_HEIGHT 128
 
 void setScreenResolution(uint16_t nw, uint16_t nh);
-void redrawScreen();
+void redrawScreen(void);
+
 #endif
