+++ ports/bare_metal_lcd_port.c
@@ -1,0 +1,13 @@
+#include "ge0_port_interface.h"
+
+void am_display_fillScreen(uint16_t color){
+
+}
+
+void am_display_drawLine(uint32_t line, uint32_t start, uint32_t width, uint32_t *colors){
+
+}
+
+void am_display_sync(void){
+
+}
